+++ dakara_player_vlc/tests/tests_dakara_server.py
