+++ dakara_player_vlc/tests/tests_font_loader.py
