+++ dakara_player_vlc/tests/tests_text_generator.py
