+++ dakara_player_vlc/tests/tests_vlc_player.py
